+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 # Unreleased
 
+- feature: re-sponsor signed forward requests
+- fix: export `get_forwarder` and `get_meta_box`
 - feature: support MetaTxRequest (user-driven, sponsor-allowed)
 - feature: support ForwardCall (type 0 payments)
 - refactor: move all rpc requests into modules